+++ src/containers/Home/Motto/Motto.tsx
@@ -1,11 +1,137 @@
 import React, { FC } from 'react'
-import Uploader from 'src/components/Uploader/Uploader'
+import { useQuery, useMutation } from '@apollo/react-hooks'
+import { useSnackbar } from 'notistack'
+import { sortBy } from 'yancey-js-util'
+import {
+  MOTTOS,
+  CREATE_ONE_MOTTO,
+  UPDATE_ONE_MOTTO,
+  DELETE_ONE_MOTTO,
+  BATCH_DELETE_MOTTO,
+  EXCHANGE_POSITION,
+} from './typeDefs'
+import { IMotto, Query } from './types'
+import MottoTable from './components/MottoTable'
+import MottoModal from './components/MottoModal'
 
 const Motto: FC = () => {
-  const onChange = (data: any) => {
-    console.log(data)
-  }
-  return <Uploader onChange={onChange} />
+  const { enqueueSnackbar } = useSnackbar()
+
+  const { loading: isFetching, data } = useQuery<Query>(MOTTOS, {
+    notifyOnNetworkStatusChange: true,
+  })
+
+  const [createMotto] = useMutation(CREATE_ONE_MOTTO, {
+    errorPolicy: 'all',
+    update(proxy, { data: { createMotto } }) {
+      const data = proxy.readQuery<Query>({ query: MOTTOS })
+
+      if (data) {
+        proxy.writeQuery({
+          query: MOTTOS,
+          data: {
+            ...data,
+            getMottos: [createMotto, ...data.getMottos],
+          },
+        })
+      }
+    },
+
+    onCompleted() {
+      enqueueSnackbar('Create success!', { variant: 'success' })
+    },
+    onError() {},
+  })
+
+  const [updateMottoById] = useMutation(UPDATE_ONE_MOTTO, {
+    errorPolicy: 'all',
+    onCompleted() {
+      enqueueSnackbar('Update success!', { variant: 'success' })
+    },
+    onError() {},
+  })
+
+  const [exchangePosition, { loading: isExchanging }] = useMutation(
+    EXCHANGE_POSITION,
+    {
+      errorPolicy: 'all',
+      onCompleted() {
+        enqueueSnackbar('Update success!', { variant: 'success' })
+      },
+      onError() {},
+    },
+  )
+
+  const [deleteMottoById, { loading: isDeleting }] = useMutation(
+    DELETE_ONE_MOTTO,
+    {
+      errorPolicy: 'all',
+      update(proxy, { data: { deleteMottoById } }) {
+        const data = proxy.readQuery<Query>({ query: MOTTOS })
+
+        if (data) {
+          proxy.writeQuery({
+            query: MOTTOS,
+            data: {
+              getMottos: data.getMottos.filter(
+                (announcement: IMotto) =>
+                  announcement._id !== deleteMottoById._id,
+              ),
+            },
+          })
+        }
+      },
+      onCompleted() {
+        enqueueSnackbar('Delete success!', { variant: 'success' })
+      },
+      onError() {},
+    },
+  )
+
+  const [deleteMottos, { loading: isBatchDeleting }] = useMutation(
+    BATCH_DELETE_MOTTO,
+    {
+      errorPolicy: 'all',
+      update(proxy, { data: { deleteMottos } }) {
+        const data = proxy.readQuery<Query>({ query: MOTTOS })
+
+        if (data) {
+          proxy.writeQuery({
+            query: MOTTOS,
+            data: {
+              getMottos: data.getMottos.filter(
+                (announcement: IMotto) =>
+                  !deleteMottos.ids.includes(announcement._id),
+              ),
+            },
+          })
+        }
+      },
+      onCompleted() {
+        enqueueSnackbar('Delete success!', { variant: 'success' })
+      },
+      onError() {},
+    },
+  )
+
+  return (
+    <>
+      <MottoTable
+        dataSource={
+          data ? data.getMottos.sort(sortBy('weight', 'descend')) : []
+        }
+        isFetching={isFetching}
+        isDeleting={isDeleting}
+        isExchanging={isExchanging}
+        isBatchDeleting={isBatchDeleting}
+        deleteMottoById={deleteMottoById}
+        deleteMottos={deleteMottos}
+        exchangePosition={exchangePosition}
+      />
+
+      <MottoModal createMotto={createMotto} updateMottoById={updateMottoById} />
+    </>
+  )
 }
 
 export default Motto